add functions CRUD api & model
this simply plumbs the function API model and CRUD endpoints, without doing
any removal of existing constructs in fn. that is, it is possible to
create/read/update/delete functions without the ability to invoke them at all,
whatsoever. the hope is that this reiterates that we have too many test
suites, reiterates that my karma is far more negative than originally thought,
and this allows the CLI and other datastore implementers to get unblocked on
this, and we can then begin the re-plumbing of agent to invoke functions
rather than routes on top of this.

note that this does not reflect extremely recent discussions around certain
fields on a func, it should be rather easy to modify this code to change/add
any fields that we want to. we need to put them on an app as we've decided,
which is pretty straightforward from this, too. The other thing that sticks
out is that PUT behaves very much like PATCH and there is maybe a discussion
to have around that (esp. keeping in mind that we're adding revisions).